+++ CHANGELOG.md
@@ -250,6 +250,7 @@
 * StatusPanelController will now (generally) try substitution lists for terms
 * StatusPanelController now properly supports all builtin gender values
 * Added RecoilEffectScale/ADSRecoilEffectScale to ranged weapon item models
+* Added NavigationNodeEx script [Experimental]
 * Fixed ThemeableElement and NonThemeableElement being ignored on child components of complex elements [Untested]
 * Fixed dialogue reactivating hidden objects even if they were initially inactive [Untested]
 * Added MaxChargeFraction to ShieldParams [Untested]